Selasa 19 November 2019

+++ Application/Controllers/AuthController.php
@@ -8,6 +8,7 @@ use Application\Foundations\QueryBuilder;
 use Application\Foundations\MailBuilder;
 use Application\Models\User;
 use Application\Models\UserConfirmation;
+use Application\Models\UserChange;
 
 class AuthController {
     public function __construct() {
@@ -17,7 +18,7 @@ class AuthController {
         if(ServiceContainer::Authentication()->isLoggedIn()) {
             return $response->redirect('/');
         }
-        return $response->view('sign-up');
+        return $response->view('auth/sign-up');
     }
     public function create_user(Request $request, Response $response) {
         if($request->email == "") {
@@ -45,6 +46,14 @@ class AuthController {
                 [ 'errors' => 'Password and confirm password must be the same' ]
             );
         }
+        $query = new QueryBuilder();
+        $query = $query->select('id')->from('user')->where('username',$request->username)->build();
+        $result = ServiceContainer::Database()->select($query);
+        if(count($result) > 0) {
+            return $response->redirect("/signup")->with( 
+                [ 'errors' => 'Username is already taken' ]
+            );
+        }
         ServiceContainer::Session()->unset('errors');
         $data = User::create([ 
             'id' => null,
@@ -55,6 +64,7 @@ class AuthController {
             'avatar_path' => '',
             'password' => password_hash($request->password,PASSWORD_DEFAULT),
             'is_confirmed' => 0,
+            'role' => 0,
             'created_at' => date('Y-m-d H:i:s'),
             'updated_at' => date('Y-m-d H:i:s'), 
         ]);
@@ -72,19 +82,17 @@ class AuthController {
            $email->from("metaforums@nanao.moe")->to($data->email)->subject("Complete your registration on Metaforums")->body($body);
            ServiceContainer::Email()->send($email);
            if($confirmator != null) {
-               return $response->redirect('/signup/success')->with(
+               return $response->redirect('/login')->with(
                    [ 'signup-email' => $request->email ],
                );
            }
         }
     }
-    public function sign_up_success(Request $request, Response $response) {
-        return $response->view('sign-up-success');
-    }
     public function sign_up_confirm(Request $request, Response $response) {
         $confirm = UserConfirmation::find($request->queryString());
-        if(isset($confirm)) {
+        if(isset($confirm) && strtotime($confirm->best_before) > time() ) {
             $id = $confirm->user_id;
+            
             $user = User::find($id);
             $user->update([ 'is_confirmed' => 1 ]);
             $confirm->delete();
@@ -95,39 +103,102 @@ class AuthController {
         if(ServiceContainer::Authentication()->isLoggedIn()) {
             return $response->redirect('/');
         }
-        return $response->view('login');
+        return $response->view('auth/login');
     }
     public function login_check(Request $request, Response $response) {
-        if ($request->username == "") {
-            return $response->redirect("/login")->with( 
-                [ 'errors' => 'Username must not be empty' ]
-            );
-        } else if ($request->password == "") {
-            return $response->redirect("/login")->with( 
-                [ 'errors' => 'Password must not be empty' ]
-            );
-        }
         $query = new QueryBuilder();
-        $query = $query->select('id,password')->from('user')->where('username',$request->username)->where('is_confirmed',1)->build();
-        $result = ServiceContainer::Database()->select($query);
-        if(count($result) == 0) {
-            return $response->redirect("/login")->with(
-                [ 'errors' => 'Wrong username or password' ]
-            );
+        $query = $query->where('username',$request->username)->orWhere('email',$request->username);
+        $result = User::selectOne($query);
+        if($result == null) {
+            if(filter_var($request->username,FILTER_VALIDATE_EMAIL)) {
+                return $response->redirect("/login")->with(
+                    [ 'errors' => 'Email is not associated with an account' ]
+                );
+            } else {
+                return $response->redirect("/login")->with(
+                    [ 'errors' => 'Username does not exist' ]
+                );
+            }
         } else {
-            $password = $result[0]["password"];
+            $password = $result->password;
             $verify = password_verify($request->password,$password);
             if(!$verify) {
                 return $response->redirect("/login")->with(
-                    [ 'errors' => 'Wrong username or password' ]
+                    [ 'errors' => 'Invalid password' ]
                 );
             }
         }
-        ServiceContainer::Session()->set('user_id',$result[0]['id']);
+        $result->update([ 'logged_in' => 1, 'last_login' => date('Y-m-d H:i:s') ]);
+        ServiceContainer::Session()->unset('errors');
+        ServiceContainer::Session()->set('user_id',$result->id);
         return $response->redirect('/');
     }
     public function logout(Request $request, Response $response) {
+        $user = User::find(ServiceContainer::Session()->get('user_id'));
+        $user->update([ 'logged_in' => 0]);
         ServiceContainer::Session()->destroy();
         return $response->redirect('/login');
+    }
+    public function forget_password(Request $request, Response $response) {
+        if(ServiceContainer::Authentication()->isLoggedIn()) {
+            return $response->redirect("/");
+        }
+        return $response->view('auth/forgot');
+    }
+    public function forget_password_confirm(Request $request, Response $response) {
+        $query = new QueryBuilder();
+        $query = $query->select('id,username,email')->from('user')->where('email',$request->email)->where('is_confirmed',1)->build();
+        $result = ServiceContainer::Database()->select($query);
+        if(count($result) > 0) {
+           $confirmator = UserChange::create([
+             'user_id' => $result[0]["id"],
+             'action_type' => 'password_reset',
+             'confirm_key' => hash('sha256',$result[0]['username'].time()),
+             'best_before' => date('Y-m-d H:i:s', strtotime('+6 hours', time())),
+           ]);
+           $email = new MailBuilder();
+           $body = "I heard you forgot your password.\n";
+           $body .= "To reset your password, use the URL below:\n\n";
+           $body .= $_SERVER['REQUEST_SCHEME'] . '://' . $_SERVER['SERVER_NAME'].'/login/reset?'.$confirmator->confirm_key;
+           $email->from("metaforums@nanao.moe")->to($result[0]['email'])->subject("Someone asked to reset your password")->body($body);
+           ServiceContainer::Email()->send($email);
+        }
+        return $response->redirect("/login/forget")->with([ 'forget_message' => 'We have sent a reset password link to the provided e-mail. If there is an account associated with the e-mail, the e-mail will be received in the inbox.' ]);
+    }
+    public function reset_password(Request $request, Response $response) {
+        if(ServiceContainer::Authentication()->isLoggedIn()) {
+            return $response->redirect("/");
+        }
+        $where = new QueryBuilder();
+        $where->where('confirm_key',$request->queryString())->where('action_type','password_reset')->where("best_before",">",date('Y-m-d H:i:s'));
+        $confirmator = UserChange::selectOne($where);
+        if(!isset($confirmator)) {
+            return $response->redirect("/");
+        }
+        return $response->view('auth/reset', [ 'key' => $request->queryString() ]);
+    }
+    public function reset_password_confirm(Request $request, Response $response) {
+        if(ServiceContainer::Authentication()->isLoggedIn()) {
+            return $response->redirect("/");
+        }
+        $where = new QueryBuilder();
+        $where->where('confirm_key',$request->confirm_key)->where('action_type','password_reset')->where("best_before",">",date('Y-m-d H:i:s'));
+        $confirmator = UserChange::selectOne($where);
+        if(!isset($confirmator)) {
+            return $response->redirect("/");
+        }
+        if (strlen($request->password) < 8) {
+            return $response->redirect("/login/reset?".$request->confirm_key)->with( 
+                [ 'errors' => 'Password must be at least 8 characters' ]
+            );
+        } else if ($request->password != $request->confirmpassword) {
+            return $response->redirect("/login/reset?".$request->confirm_key)->with( 
+                [ 'errors' => 'Password and confirm password must be the same' ]
+            );
+        }
+        $user = User::find($confirmator->user_id);
+        $user->update(['password' => password_hash($request->password,PASSWORD_DEFAULT) ]);
+        $confirmator->delete();
+        return $response->redirect("/login");
     }
 }
